fix: extract out enrichUser function
