+++ src/routes/management-api/users-by-email.ts
@@ -11,42 +11,12 @@ import {
 import { RequestWithContext } from "../../types/RequestWithContext";
 import { HTTPException } from "hono/http-exception";
 import { UserResponse } from "../../types/auth0/UserResponse";
-import userIdParse from "../../utils/userIdParse";
-import { Env, User } from "../../types";
+import { enrichUser } from "../../utils/enrichUser";
 
 export interface LinkBodyParams {
   provider?: string;
   connection_id?: string;
   link_with: string;
-}
-
-async function enrichUser(
-  env: Env,
-  tenantId: string,
-  primaryUser: User,
-): Promise<UserResponse> {
-  const linkedusers = await env.data.users.list(tenantId, {
-    page: 0,
-    per_page: 10,
-    include_totals: false,
-    q: `linked_to:${primaryUser.id}`,
-  });
-
-  const identities = [primaryUser, ...linkedusers.users].map((u) => ({
-    connection: u.connection,
-    provider: u.provider,
-    user_id: userIdParse(u.id),
-    isSocial: u.is_social,
-    // TODO - have profileData PR that needs continuing...
-  }));
-
-  const { id, ...userWithoutId } = primaryUser;
-
-  return {
-    ...userWithoutId,
-    identities,
-    user_id: primaryUser.id,
-  };
 }
 
 @Route("api/v2/users-by-email")
